Enable reuse of indices across stochastic objective evaluations
Needed for GCP-Adam.

Alt names for the `SampleOnce` struct
+ FixedSampler
+ FrozenSampler
+ SampleThenFreeze

Also considered making it a mutable struct
so we can assign new arrays to `cache` field.
Would also be able to have a `already_sampled`
Bool field to explicitly indicate if we have sampled
already or not (currently handled implicitly by
checking if the cache is empty).

Also considered providing this functionality
by implementing a `ConstantSampler` type
that just holds indices, values, and weights.